+++ benchmarks/vector-algorithms-0.5.4.2/Data/Vector/Algorithms/Termination.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
-{-@ LIQUID "--real" @-}
 module Term where
 
 import Data.Vector.Algorithms.Common (shiftRI)