+++ src/Language/Haskell/Liquid/Bare/Check.hs
@@ -237,7 +237,6 @@ checkRType :: (PPrint r, Reftable r) => TCEmb TyCon -> SEnv SortedReft -> RRType
 ------------------------------------------------------------------------------------------------
 
 checkRType emb env t   =  checkAppTys t
-                      <|> checkFunRefs t
                       <|> checkAbstractRefs t
                       <|> efoldReft cb (rTypeSortedReft emb) f insertPEnv env Nothing t
   where
@@ -274,7 +273,7 @@ checkTcArity (RTyCon { rtc_tc = tc }) givenArity
   where
     expectedArity = realTcArity tc
 
-
+{- 
 checkFunRefs t = go t
   where
     go (RAllT _ t)      = go t
@@ -291,6 +290,7 @@ checkFunRefs t = go t
     go (RFun _ t1 t2 r)
       | isTauto r       = go t1 <|> go t2
       | otherwise       = Just $ text "Function types cannot have refinements:" <+> (pprint r)
+-} 
 
 checkAbstractRefs t = go t
   where