+++ src/Language/Haskell/Liquid/Bare/Resolve.hs
@@ -59,6 +59,7 @@ instance Resolvable Expr where
   resolve l (PImp p q)      = PImp    <$> resolve l p  <*> resolve l q
   resolve l (PIff p q)      = PIff    <$> resolve l p  <*> resolve l q
   resolve l (PAtom r e1 e2) = PAtom r <$> resolve l e1 <*> resolve l e2
+  resolve l (ELam (x,t) e)  = ELam    <$> ((,) <$> resolve l x <*> resolve l t) <*> resolve l e
   resolve l (PAll vs p)     = PAll    <$> mapM (secondM (resolve l)) vs <*> resolve l p
   resolve l (ETApp e s)     = ETApp   <$> resolve l e <*> resolve l s 
   resolve l (ETAbs e s)     = ETAbs   <$> resolve l e <*> resolve l s 
@@ -66,6 +67,7 @@ instance Resolvable Expr where
   resolve l (PExist ss e)   = PExist ss <$> resolve l e
   resolve _ (ESym s)        = return $ ESym s 
   resolve _ (ECon c)        = return $ ECon c 
+  resolve _ PGrad           = return PGrad 
 
 instance Resolvable LocSymbol where
   resolve _ ls@(Loc l l' s)