+++ src/Language/Haskell/Liquid/Constraint/ToFixpoint.hs
@@ -25,7 +25,7 @@ cgInfoFInfo info cgi fi = do
   return    $ tgtFI <> impFI
 
 targetFInfo :: GhcInfo -> CGInfo -> FilePath -> F.FInfo Cinfo
-targetFInfo info cgi fn = F.fi cs ws bs ls ks qs bi fn
+targetFInfo info cgi fn = F.fi cs ws bs ls ks qs bi fn aHO 
   where
    cs     = fixCs  cgi
    ws     = fixWfs cgi
@@ -34,6 +34,7 @@ targetFInfo info cgi fn = F.fi cs ws bs ls ks qs bi fn
    ks     = kuts cgi
    qs     = targetQuals info cgi
    bi     = (`Ci` Nothing) <$> bindSpans cgi
+   aHO    = allowHO cgi 
 
 targetQuals :: GhcInfo -> CGInfo -> [F.Qualifier]
 targetQuals info cgi = spcQs ++ genQs