+++ src/Language/Haskell/Liquid/Types.hs
@@ -58,7 +58,6 @@ module Language.Haskell.Liquid.Types (
 
   -- * Classes describing operations on `RTypes`
   , TyConable (..)
-  , RefTypable (..)
   , SubsTy (..)
 
   -- * Predicate Variables
@@ -260,6 +259,7 @@ data PPEnv
        , ppSs    :: Bool
        , ppShort :: Bool
        }
+    deriving (Show)
 
 ppEnv           = ppEnvPrintPreds
 _ppEnvCurrent   = PP False False False False
@@ -355,10 +355,20 @@ toLogicMap ls = mempty {logic_map = M.fromList $ map toLMap ls}
     toLMap (x, xs, e) = (x, LMap {lvar = x, largs = xs, lexpr = e})
 
 eAppWithMap lmap f es def
-  | Just (LMap _ xs e) <- M.lookup (val f) (logic_map lmap)
+  | Just (LMap _ xs e) <- M.lookup (val f) (logic_map lmap), length xs == length es
   = subst (mkSubst $ zip xs es) e
+  | Just (LMap _ xs e) <- M.lookup (val f) (logic_map lmap), isApp e
+  = subst (mkSubst $ zip xs es) $ dropApp e (length xs - length es)
   | otherwise
   = def
+
+dropApp e i | i <= 0 = e
+dropApp (EApp e _) i = dropApp e (i-1)
+dropApp _ _          = errorstar "impossible"
+
+isApp (EApp (EVar _) (EVar _)) = True
+isApp (EApp e (EVar _))        = isApp e
+isApp _                        = False
 
 data TyConP = TyConP { freeTyVarsTy :: ![RTyVar]
                      , freePredTy   :: ![PVar RSort]
@@ -710,17 +720,6 @@ class (Eq c) => TyConable c where
   isNumCls  = const False
   isFracCls = const False
 
-class ( TyConable c
-      , Eq c, Eq tv
-      , Hashable tv
-      , Reftable r
-      , PPrint r
-      ) => RefTypable c tv r
-  where
-    ppRType  :: Prec -> RType c tv r -> Doc
-
-
-
 -------------------------------------------------------------------------------
 -- | TyConable Instances -------------------------------------------------------
 -------------------------------------------------------------------------------
@@ -755,13 +754,13 @@ instance Eq RTyCon where
   x == y = rtc_tc x == rtc_tc y
 
 instance Fixpoint RTyCon where
-  toFix (RTyCon c _ _) = text $ showPpr c -- <+> text "\n<<" <+> hsep (map toFix ts) <+> text ">>\n"
+  toFix (RTyCon c _ _) = text $ showPpr c
 
 instance Fixpoint Cinfo where
   toFix = text . showPpr . ci_loc
 
 instance PPrint RTyCon where
-  pprint = text . showPpr . rtc_tc
+  pprintTidy _ = text . showPpr . rtc_tc
 
 
 instance Show RTyCon where
@@ -1005,7 +1004,7 @@ instance Subable r => Subable (UReft r) where
   substf f (MkUReft r z l) = MkUReft (substf f r) (substf f z) (substf f l)
   substa f (MkUReft r z l) = MkUReft (substa f r) (substa f z) (substa f l)
 
-instance (Reftable r, RefTypable c tv r) => Subable (RTProp c tv r) where
+instance (Reftable r, TyConable c) => Subable (RTProp c tv r) where
   syms (RProp  ss r)     = (fst <$> ss) ++ syms r
 
   subst su (RProp ss (RHole r)) = RProp ss (RHole (subst su r))
@@ -1018,15 +1017,12 @@ instance (Reftable r, RefTypable c tv r) => Subable (RTProp c tv r) where
   substa f (RProp  ss t) = RProp ss (substa f <$> t)
 
 
-instance (Subable r, RefTypable c tv r) => Subable (RType c tv r) where
+instance (Subable r, Reftable r, TyConable c) => Subable (RType c tv r) where
   syms        = foldReft (\_ r acc -> syms r ++ acc) []
   substa f    = mapReft (substa f)
   substf f    = emapReft (substf . substfExcept f) []
   subst su    = emapReft (subst  . substExcept su) []
   subst1 t su = emapReft (\xs r -> subst1Except xs r su) [] t
-
-
-
 
 instance Reftable Predicate where
   isTauto (Pr ps)      = null ps
@@ -1237,6 +1233,7 @@ rTypeValueVar t = vv where Reft (vv,_) =  rTypeReft t
 rTypeReft :: (Reftable r) => RType c tv r -> Reft
 rTypeReft = fromMaybe trueReft . fmap toReft . stripRTypeBase
 
+
 -- stripRTypeBase ::  RType a -> Maybe a
 stripRTypeBase (RApp _ _ _ x)
   = Just x
@@ -1279,14 +1276,14 @@ instance Show Stratum where
   show (SVar s) = show s
 
 instance PPrint Stratum where
-  pprint = text . show
+  pprintTidy _ = text . show
 
 instance PPrint Strata where
-  pprint [] = empty
-  pprint ss = hsep (pprint <$> nub ss)
+  pprintTidy _ [] = empty
+  pprintTidy k ss = hsep (pprintTidy k <$> nub ss)
 
 instance PPrint (PVar a) where
-  pprint = ppr_pvar
+  pprintTidy _ = ppr_pvar
 
 ppr_pvar :: PVar a -> Doc
 ppr_pvar (PV s _ _ xts) = pprint s <+> hsep (pprint <$> dargs xts)
@@ -1295,13 +1292,24 @@ ppr_pvar (PV s _ _ xts) = pprint s <+> hsep (pprint <$> dargs xts)
 
 
 instance PPrint Predicate where
-  pprint (Pr [])       = text "True"
-  pprint (Pr pvs)      = hsep $ punctuate (text "&") (map pprint pvs)
+  pprintTidy _ (Pr [])  = text "True"
+  pprintTidy k (Pr pvs) = hsep $ punctuate (text "&") (pprintTidy k <$> pvs)
 
 
--- | The type used during constraint generation, used also to define contexts
--- for errors, hence in this file, and NOT in Constraint.hs
-newtype REnv = REnv  (M.HashMap Symbol SpecType)
+-- | The type used during constraint generation, used
+--   also to define contexts for errors, hence in this
+--   file, and NOT in elsewhere. **DO NOT ATTEMPT TO MOVE**
+--   Am splitting into
+--   + global : many bindings, shared across all constraints
+--   + local  : few bindings, relevant to particular constraints
+
+data REnv = REnv
+  { reGlobal :: M.HashMap Symbol SpecType -- ^ the "global" names for module
+  , reLocal  :: M.HashMap Symbol SpecType -- ^ the "local" names for sub-exprs
+  }
+
+instance NFData REnv where
+  rnf (REnv {}) = ()
 
 ------------------------------------------------------------------------
 -- | Error Data Type ---------------------------------------------------
@@ -1401,23 +1409,25 @@ data CMeasure ty = CM
   } deriving (Data, Typeable, Generic, Functor)
 
 instance PPrint Body where
-  pprint (E e)   = pprint e
-  pprint (P p)   = pprint p
-  pprint (R v p) = braces (pprint v <+> text "|" <+> pprint p)
+  pprintTidy k (E e)   = pprintTidy k e
+  pprintTidy k (P p)   = pprintTidy k p
+  pprintTidy k (R v p) = braces (pprintTidy k v <+> "|" <+> pprintTidy k p)
 
 instance PPrint a => PPrint (Def t a) where
-  pprint (Def m p c _ bs body) = pprint m <+> pprint (fst <$> p) <+> cbsd <> text " = " <> pprint body
-    where cbsd = parens (pprint c <> hsep (pprint `fmap` (fst <$> bs)))
+  pprintTidy k (Def m p c _ bs body)
+           = pprintTidy k m <+> pprintTidy k (fst <$> p) <+> cbsd <+> "=" <+> pprintTidy k body
+    where
+      cbsd = parens (pprintTidy k c <> hsep (pprintTidy k `fmap` (fst <$> bs)))
 
 instance (PPrint t, PPrint a) => PPrint (Measure t a) where
-  pprint (M n s eqs) =  pprint n <> text " :: " <> pprint s
-                     $$ vcat (pprint `fmap` eqs)
+  pprintTidy k (M n s eqs) =  pprintTidy k n <+> "::" <+> pprintTidy k s
+                              $$ vcat (pprintTidy k `fmap` eqs)
 
 instance PPrint (Measure t a) => Show (Measure t a) where
   show = showpp
 
 instance PPrint t => PPrint (CMeasure t) where
-  pprint (CM n s) =  pprint n <> text " :: " <> pprint s
+  pprintTidy k (CM n s) =  pprintTidy k n <+> "::" <+> pprintTidy k s
 
 instance PPrint (CMeasure t) => Show (CMeasure t) where
   show = showpp
@@ -1540,10 +1550,10 @@ updKVProf k kvs (KVP m) = KVP $ M.insert k (kn + n) m
 instance NFData KVKind
 
 instance PPrint KVKind where
-  pprint = text . show
+  pprintTidy _ = text . show
 
 instance PPrint KVProf where
-  pprint (KVP m) = pprint $ M.toList m
+  pprintTidy _ (KVP m) = pprint $ M.toList m
 
 instance NFData KVProf
 
@@ -1566,7 +1576,7 @@ instance Symbolic DataCon where
 
 
 instance PPrint DataCon where
-  pprint = text . showPpr
+  pprintTidy _ = text . showPpr
 
 instance Show DataCon where
   show = showpp
@@ -1593,7 +1603,7 @@ instance Bifunctor MSpec   where
   second                    = fmap
 
 instance (PPrint t, PPrint a) => PPrint (MSpec t a) where
-  pprint =  vcat . fmap pprint . fmap snd . M.toList . measMap
+  pprintTidy k =  vcat . fmap (pprintTidy k) . fmap snd . M.toList . measMap
 
 instance (Show ty, Show ctor, PPrint ctor, PPrint ty) => Show (MSpec ty ctor) where
   show (MSpec ct m cm im)
@@ -1621,21 +1631,21 @@ instance Eq ctor => Monoid (MSpec ty ctor) where
 --------------------------------------------------------------------------------
 
 instance PPrint RTyVar where
-  pprint (RTV α)
-   | ppTyVar ppEnv = ppr_tyvar α
-   | otherwise     = ppr_tyvar_short α
-
-ppr_tyvar       = text . tvId
-ppr_tyvar_short = text . showPpr
+  pprintTidy _ (RTV α)
+   | ppTyVar ppEnv  = ppr_tyvar α
+   | otherwise      = ppr_tyvar_short α
+   where
+    ppr_tyvar       = text . tvId
+    ppr_tyvar_short = text . showPpr
 
 instance (PPrint r, Reftable r, PPrint t, PPrint (RType c tv r)) => PPrint (Ref t (RType c tv r)) where
-  pprint (RProp ss (RHole s)) = ppRefArgs (fst <$> ss) <+> pprint s
-  pprint (RProp ss s) = ppRefArgs (fst <$> ss) <+> pprint (fromMaybe mempty (stripRTypeBase s))
+  pprintTidy k (RProp ss (RHole s)) = ppRefArgs k (fst <$> ss) <+> pprintTidy k s
+  pprintTidy k (RProp ss s)         = ppRefArgs k (fst <$> ss) <+> pprintTidy k (fromMaybe mempty (stripRTypeBase s))
 
 
-ppRefArgs :: [Symbol] -> Doc
-ppRefArgs [] = empty
-ppRefArgs ss = text "\\" <> hsep (ppRefSym <$> ss ++ [vv Nothing]) <+> text "->"
+ppRefArgs :: Tidy -> [Symbol] -> Doc
+ppRefArgs _ [] = empty
+ppRefArgs k ss = text "\\" <> hsep (ppRefSym k <$> ss ++ [vv Nothing]) <+> "->"
 
-ppRefSym "" = text "_"
-ppRefSym s  = pprint s
+ppRefSym _ "" = text "_"
+ppRefSym k s  = pprintTidy k s