+++ src/Language/Haskell/Liquid/Types/Errors.hs
@@ -52,7 +52,7 @@ import           Data.Maybe
 import           Text.PrettyPrint.HughesPJ
 import           Data.Aeson hiding (Result)
 import qualified Data.HashMap.Strict as M
-import           Language.Fixpoint.Types      (showpp, Tidy (..), PPrint (..), Symbol, Expr, Reft)
+import           Language.Fixpoint.Types      (pprint, showpp, Tidy (..), PPrint (..), Symbol, Expr, Reft)
 import           Language.Fixpoint.Misc (dcolon)
 import           Language.Haskell.Liquid.Misc (intToString)
 import           Text.Parsec.Error            (ParseError)
@@ -66,11 +66,11 @@ import           Text.Parsec.Error (ParseError, errorMessages, showErrorMessages
 import           GHC.Stack
 
 instance PPrint ParseError where
-  pprint e = vcat $ tail $ map text ls
+  pprintTidy _ e = vcat $ tail $ text <$> ls
     where
-      ls = lines $ showErrorMessages "or" "unknown parse error"
-                                     "expecting" "unexpected" "end of input"
-                                     (errorMessages e)
+      ls         = lines $ showErrorMessages "or" "unknown parse error"
+                               "expecting" "unexpected" "end of input"
+                               (errorMessages e)
 
 --------------------------------------------------------------------------------
 -- | Context information for Error Messages ------------------------------------
@@ -147,7 +147,7 @@ instance Show Oblig where
 instance NFData Oblig
 
 instance PPrint Oblig where
-  pprint = ppOblig
+  pprintTidy _ = ppOblig
 
 ppOblig :: Oblig -> Doc
 ppOblig OCons = text "Constraint Check"
@@ -231,8 +231,8 @@ data TError t =
                 } -- ^ Using  sort error
 
   | ErrIAlMis   { pos :: !SrcSpan
-                , t1  :: !t
-                , t2  :: !t
+                , tAs :: !t
+                , tUs :: !t
                 , msg :: !Doc
                 } -- ^ Incompatible using error
 
@@ -308,7 +308,7 @@ data TError t =
                 , msg   :: !Doc
                 } -- ^ Sigh. Other.
 
-  deriving (Typeable, Generic, Functor)
+  deriving (Typeable, Generic , Functor )
 
 instance NFData ParseError where
   rnf t = seq t ()
@@ -339,7 +339,7 @@ instance Ex.Error (TError a) where
 type UserError  = TError Doc
 
 instance PPrint SrcSpan where
-  pprint = text . showSDoc . Out.ppr
+  pprintTidy _ = text . showSDoc . Out.ppr
      where
         showSDoc sdoc = Out.renderWithStyle
                         unsafeGlobalDynFlags
@@ -348,8 +348,7 @@ instance PPrint SrcSpan where
                               Out.AllTheWay)
 
 instance PPrint UserError where
-  pprint       = pprintTidy Full
-  pprintTidy k = ppError k empty . fmap pprint
+  pprintTidy k = ppError k empty . fmap (pprintTidy Lossy)
 
 instance Show UserError where
   show = showpp
@@ -410,40 +409,24 @@ ppFull :: Tidy -> Doc -> Doc
 ppFull Full  d = d
 ppFull Lossy _ = empty
 
-ppReqInContext :: (PPrint t, PPrint c) => t -> t -> c -> Doc
-ppReqInContext tA tE c
+ppReqInContext :: (PPrint t, PPrint c) => Tidy -> t -> t -> c -> Doc
+ppReqInContext _ tA tE c
   = sepVcat blankLine
       [ nests 2 [ text "Inferred type"
-                , text "VV :" <+> pprint tA]
+                , text "VV :" <+> pprintTidy Lossy tA]
       , nests 2 [ text "not a subtype of Required type"
-                , text "VV :" <+> pprint tE]
+                , text "VV :" <+> pprintTidy Lossy tE]
       , nests 2 [ text "In Context"
-                , pprint c                 ]]
+                , pprintTidy Lossy c                 ]]
 
 
-ppPropInContext :: (PPrint p, PPrint c) => p -> c -> Doc
-ppPropInContext p c
+ppPropInContext :: (PPrint p, PPrint c) => Tidy -> p -> c -> Doc
+ppPropInContext _ p c
   = sepVcat blankLine
       [ nests 2 [ text "Property"
-                , pprint p]
+                , pprintTidy Lossy p]
       , nests 2 [ text "Not provable in context"
-                , pprint c                 ]]
-
-{-
-pprintCtx :: (PTable c) => c -> Doc
-pprintCtx = pprint . ptable
-
-instance (PPrint a, PPrint b) => PTable (M.HashMap a b) where
-  ptable t = DocTable [ (pprint k, pprint v) | (k, v) <- M.toList t]
-
-
-pprintKVs :: (PPrint k, PPrint v) => [(k, v)] -> Doc
-pprintKVs = vcat . punctuate (text "\n") . map pp1
-  where
-    pp1 (x,y) = pprint x <+> text ":=" <+> pprint y
--}
-
-
+                , pprintTidy Lossy c                 ]]
 
 instance ToJSON RealSrcSpan where
   toJSON sp = object [ "filename"  .= f  -- (unpackFS $ srcSpanFile sp)
@@ -508,17 +491,17 @@ ppError' :: (PPrint a, Show a) => Tidy -> Doc -> Doc -> TError a -> Doc
 ppError' td dSp dCtx (ErrAssType _ o _ c p)
   = dSp <+> pprint o
         $+$ dCtx
-        $+$ (ppFull td $ ppPropInContext p c)
+        $+$ (ppFull td $ ppPropInContext td p c)
 
 ppError' td dSp dCtx (ErrSubType _ _ c tA tE)
   = dSp <+> text "Liquid Type Mismatch"
         $+$ dCtx
-        $+$ (ppFull td $ ppReqInContext tA tE c)
+        $+$ (ppFull td $ ppReqInContext td tA tE c)
 
 ppError' td  dSp dCtx (ErrFCrash _ _ c tA tE)
   = dSp <+> text "Fixpoint Crash on Constraint"
         $+$ dCtx
-        $+$ (ppFull td $ ppReqInContext tA tE c)
+        $+$ (ppFull td $ ppReqInContext td tA tE c)
 
 ppError' _ dSp dCtx (ErrParse _ _ e)
   = dSp <+> text "Cannot parse specification:"