+++ tests/equationalproofs/pos/MonadicLaws.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE ExtendedDefaultRules #-}
 
+{-@ LIQUID "--higherorder"      @-}
 {-@ LIQUID "--autoproofs"      @-}
 {-@ LIQUID "--totality"        @-}
 {-@ LIQUID "--exact-data-cons" @-}
@@ -22,6 +23,7 @@ $(axiomatize
       return x = C x N
     |])
 
+
 {-@ axiomatize append @-}
 $(axiomatize
   [d| append :: L a -> L a -> L a
@@ -38,15 +40,15 @@ $(axiomatize
 
 
 -- NV TODO:
--- 1. remove the manual runFun there
 -- 2. check why failure to prove takes so long
 
 -- | Left identity: return a >>= f ≡ f a
 
 prop_left_identity :: Eq a => a -> (a -> L a) -> Proof
 {-@ prop_left_identity :: x:a -> f:(a -> L a)
-                       -> {v:Proof | bind (return x) f == runFun f x} @-}
-prop_left_identity x f = undefined -- auto 2 (bind (return x) f == f x)
+                       -> {v:Proof | bind (return x) f == f x} @-}
+prop_left_identity x f = auto 2 (bind (return x) f == f x)
+{- 
   where
     e1  = bind (return x) f
     pr1 = axiom_return x
@@ -57,24 +59,12 @@ prop_left_identity x f = undefined -- auto 2 (bind (return x) f == f x)
     e4  = append (f x) N
     pr4 = prop_app_nil (f x)
     e5  = f x
+-}
 
 -- | Right Identity m >>= return ≡  m
-{-@ prop_right_identity :: Eq a => L a -> {v:Proof | bind N return == N } @-}
+{-@ prop_right_identity :: Eq a => xs:L a -> {v:Proof | bind xs return == xs } @-}
 prop_right_identity :: Eq a => L a -> Proof
-prop_right_identity N =  refl (bind N returnZZZ) `by` pr1 -- auto 1 (bind N return == N) --  `by` pr1
-  where
-    e1  = bind N return
-    pr1 = axiom_bind_N returnZZZ
-    e2  = N
-    returnZZZ = undefined
-
-prop_right_identity (C x xs) = undefined
-
-{-@ f :: x:Int -> {v:Int | v < x} @-}
-f :: Int -> Int
-f x = x
-
-
+prop_right_identity xs =  cases 2 (bind xs return == xs) 
 
 {-@ prop_app_nil :: ys:L a -> {v:Proof | append ys N == ys} @-}
 prop_app_nil :: (Eq a) => L a -> Proof