+++ CHANGES.md
@@ -5,6 +5,8 @@
 ## 0.5.0.1
 
 - Fixed a bug in the specification for `Data.Traversable.sequence`
+- Make interpreted mul and div the default, when `solver = z3`
+- Use `--higherorder` to allow higher order binders into the fixpoint environment 
 
 ## 0.5.0.0